requirements.txt
asgiref==3.9.2
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mailtrap==2.0.1
numpy==2.3.3
packaging==25.0
pandas==2.2.3
psycopg==3.2.3
psycopg-binary==3.2.3
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.27.1
six==1.17.0
sqlparse==0.5.3
stripe==12.5.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
kaggle==1.6.17
Pillow==10.4.0
opencv-python
boto3